PORTFOLIO/requirement.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
decorator==5.1.1
Deprecated==1.2.14
Django==5.1
django-livereload-server==0.5.1
django-phonenumber-field==8.0.0
exceptiongroup==1.2.2
executing==2.0.1
frozenlist==1.4.1
idna==3.7
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
multidict==6.0.5
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.43
pillow==10.4.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
watchfiles==0.23.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
